installed react-slick library for carousel; removed carousel bootsrap due to lag; added placeholder img for testing
